+++ src/Engines/EloquentEngine.php
@@ -21,7 +21,10 @@ class EloquentEngine extends QueryBuilderEngine
      */
     public function __construct($model, Request $request)
     {
-        $this->query = $model instanceof Builder ? $model : $model->getQuery();
-        $this->init($request, $this->query->getQuery(), 'eloquent');
+        $builder = $model instanceof Builder ? $model : $model->getQuery();
+        parent::__construct($builder->getQuery(), $request);
+
+        $this->query      = $builder;
+        $this->query_type = 'eloquent';
     }
 }